refactor(lenses): address comments - composable moved
